moved main class to own file + added namespacing
